fix: add utils module and refactor os cups
